Finish HW4 (guard fix, add auth with JWT, owner

+++ repository/contacts.js
@@ -1,28 +1,36 @@
 const Contact = require('../models/contacts')
 
-const listContacts = async () => {
-  const result = await  Contact.find()
+
+const listContacts = async (query, user) => {
+  const result = await  Contact.find({ owner: user._id }).populate({
+    path: 'owner',
+    select: 'name email subscription',
+  }) 
   return result
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId })     // or we can use - Contact.findById(contactId)
+const getContactById = async (contactId, user) => {
+  const result = await Contact.findOne({ _id: contactId, owner: user._id })
+  .populate({
+    path: 'owner',
+    select: 'name email subscription',
+  }) 
   return result
 }
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndRemove({ _id: contactId })
+const removeContact = async (contactId, user) => {
+  const result = await Contact.findOneAndRemove({ _id: contactId, owner: user._id })
   return result
 }
 
-const addContact = async (body) => {
-  const result = await Contact.create(body)
+const addContact = async (body, user) => {
+  const result = await Contact.create({...body, owner: user._id})
   return result
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, user) => {
   const result = await Contact.findOneAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: user._id },
     { ...body },
     { new: true },
   )
